remove has_any_labels_of
