+++ src/Application/Rule/Rules/HasAnyLabelsRule.php
@@ -4,22 +4,28 @@ namespace ArtARTs36\MergeRequestLinter\Application\Rule\Rules;
 
 use ArtARTs36\MergeRequestLinter\Application\Rule\Definition\Definition;
 use ArtARTs36\MergeRequestLinter\Domain\Request\MergeRequest;
-use ArtARTs36\MergeRequestLinter\Domain\Rule\Rule;
 use ArtARTs36\MergeRequestLinter\Domain\Rule\RuleDefinition;
 use ArtARTs36\MergeRequestLinter\Shared\Attributes\Description;
 
-#[Description('Merge Request must have any labels.')]
-final class HasAnyLabelsRule extends AbstractRule implements Rule
+#[Description('Merge Request must have any {labels}.')]
+final class HasAnyLabelsRule extends AbstractLabelsRule
 {
     public const NAME = '@mr-linter/has_any_labels';
 
     protected function doLint(MergeRequest $request): bool
     {
-        return ! $request->labels->isEmpty();
+        return ! $this->labels->diff($request->labels)->equalsCount($this->labels);
     }
 
     public function getDefinition(): RuleDefinition
     {
-        return new Definition("Merge Request must have any labels");
+        if ($this->labels->isEmpty()) {
+            return new Definition('Merge Request must have any labels');
+        }
+
+        return new Definition(sprintf(
+            'Merge Request must have any labels of: [%s]',
+            $this->labels->implode(', '),
+        ));
     }
 }
