+++ tests/Unit/Application/Rule/Rules/HasAnyLabelsRuleTest.php
@@ -4,6 +4,7 @@ namespace ArtARTs36\MergeRequestLinter\Tests\Unit\Application\Rule\Rules;
 
 use ArtARTs36\MergeRequestLinter\Application\Rule\Rules\HasAnyLabelsRule;
 use ArtARTs36\MergeRequestLinter\Domain\Request\MergeRequest;
+use ArtARTs36\MergeRequestLinter\Shared\DataStructure\Set;
 use ArtARTs36\MergeRequestLinter\Tests\TestCase;
 
 final class HasAnyLabelsRuleTest extends TestCase
@@ -12,14 +13,27 @@ final class HasAnyLabelsRuleTest extends TestCase
     {
         return [
             [
-                $this->makeMergeRequest(),
-                true,
+                $this->makeMergeRequest([
+                    'labels' => [],
+                ]),
+                ['Feature'],
+                'hasNotes' => true,
             ],
             [
                 $this->makeMergeRequest([
-                    'labels' => ['Feature'],
+                    'labels' => [],
                 ]),
-                false,
+                [],
+                'hasNotes' => true,
+            ],
+            [
+                $this->makeMergeRequest([
+                    'labels' => [
+                        'Feature',
+                    ],
+                ]),
+                ['Feature', 'Backend'],
+                'hasNotes' => false,
             ],
         ];
     }
@@ -28,10 +42,11 @@ final class HasAnyLabelsRuleTest extends TestCase
      * @dataProvider providerForTestLint
      * @covers \ArtARTs36\MergeRequestLinter\Application\Rule\Rules\HasAnyLabelsRule::lint
      * @covers \ArtARTs36\MergeRequestLinter\Application\Rule\Rules\HasAnyLabelsRule::doLint
+     * @covers \ArtARTs36\MergeRequestLinter\Application\Rule\Rules\HasAnyLabelsRule::__construct
      * @covers \ArtARTs36\MergeRequestLinter\Application\Rule\Rules\HasAnyLabelsRule::getDefinition
      */
-    public function testLint(MergeRequest $request, bool $hasNotes): void
+    public function testLint(MergeRequest $request, array $requestedLabels, bool $hasNotes): void
     {
-        self::assertHasNotes($request, new HasAnyLabelsRule(), $hasNotes);
+        self::assertHasNotes($request, new HasAnyLabelsRule(Set::fromList($requestedLabels)), $hasNotes);
     }
 }
